infrastructure_resources/requirements.txt
cdktf~=0.0.19
python-decouple
pycryptodome
Flask
gunicorn
Flask-AutoIndex
